+++ fits_storage/scripts/problem_checker.py
@@ -200,8 +200,6 @@ if __name__ == "__main__":
         DHSProblemChecker(),
     ]
 
-    print("Starting problem_checker.py")
-
     problems = list()
 
     with session_scope() as session:
@@ -219,6 +217,6 @@ if __name__ == "__main__":
             message = "From: %s\r\nTo: %s\r\nSubject: %s\r\n\r\n%s" % (
                 mailfrom, ", ".join(mailto), "ERRORS - %s" % subject, '\n'.join(problems))
 
-        server = smtplib.SMTP(smtp_server)
-        server.sendmail(mailfrom, mailto, message)
-        server.quit()
+            server = smtplib.SMTP(smtp_server)
+            server.sendmail(mailfrom, mailto, message)
+            server.quit()